refactor: relocated shared components
